fix: Remove cache cleanup from app startup
This could lead to very long startup time when the cache contains a lot
of pictures. A possible other implementation would have been to use
threading here, but tray icon does not like that, and in any case the
daemon itself already take care of calling this function when it stops.